finish monochrome bmp

+++ Hebb/BmpWriter.h
@@ -5,7 +5,7 @@
 class BmpWriter
 {
 public:
-    BmpWriter(unsigned count, double *blocks, const BITMAPHEADER& header, char *headerRest, unsigned headerRestCount);
+    BmpWriter(unsigned count, float *blocks, const BITMAPHEADER& header, char *headerRest, unsigned headerRestCount);
     ~BmpWriter();
     void Write(const char *bmpPath)const;
 private:
